<article> tag addition
i added a main <article> addition (After studying Vitali's notes for his written exam, I saw he mentioned this multiple times, how this tag should be used for things that will be styled similarly, like blog posts or as in our case. This also makes it easier for us to work with the 'main' body of the html pages as a whole, excluding the header and footer. As far as I can tell, adding it doesn't bring any problems but allows us to do much more. If you disagree, we can discuss it in person more!